Tap2Unlock - a new screen wake feature similar to d2w and s2w
Divide your screen vertically and horizontally
They are numbered as
          1 | 2
          -----
          3 | 4
The pattern is stored in sysfs file sys/android_touch/tap2unlock_pattern
You tap the right combination to open your screen
Pressing the power button will turn your display on but not the touch
sensor.

version 1.0
-Ability - 4 number pattern (will be improved)
-Currently the pattern gets rstored to 1234 after every boot
(This can only be fixed if a kernel mod app provides support for it in
the future.)

All copyrights of t2u is reserved with goutamniwas
<goutamniwas@gmail.com>
Many credits to showp-1984 for his D2w which made this possible.

Note: This was ported from the original commits listed at
http://forum.xda-developers.com/android/software-hacking/kernel-tap2unlo
ck-tap2wake-feature-t2965344 but is poorly written and will most likely
break. This is only an attempt to salvage the concept

Conflicts:

	drivers/input/touchscreen/Makefile
	drivers/video/msm/mdss/mdss_dsi_panel.c