T2U : Prevent touch freeze during phonecalls
The screen eventhough will not freeze when powered off by proximity
sensor which was due to a previous commmit .
But incase when the screen powers off due to timeout , which happens
when you speak in loudspeaker , the wake up doesnt resume your touch
sensor.
So this commit solves the issue , v2.0 allows the variable t2u_allow to
be controlled with sysfs (only when u know the pattern)
with that the t2u application is able to make the touch sensor wake or
freeze during calls ,we make use of it to set a new state variable
incall_active which sets to true if t2u_allow is set to 1 by the
application .
And when its set to 0 , the screen freezes .
So when call active the application controls the touch sensor , this
commit enables or allows the app to do so.

Also the commit also has elements to allow freeze even during call
active , when the power off is due to proximity , as ther is no use of
proximity if the touch can never freeze during call.